Add a package for ViewModel
It took me a bit to get it figured out. The main trick seems to be adding the package as a dependency for the main application(in this case Matrix Calculator). For some reason this affected the ability of the package to see itself as a module in its own tests. I don't quite understand that but maybe someday I will.