Rename Application to VMDLWindow
In iOS there's only one window but in macOS there can be more than one so it's not quite correct to say that code managing the main(and only) MatrixCalc window is the application. I also didn't like how Application was get confused with other Applications in the Apple libraries. There does seem to be something clever that avoids confusion but it's not quite what I wanted.