Metal003 Setup ALL the bits to get make the MTKViewDelegate work
Ok, this was quite a bit of stuff. I should've done it in smaller bits. Once I started pushing towards getting the MTKViewDelegate working I had to put a whole bunch of stuff in place. Next time, I think I'll see if I can build towards it more. I think I may only need the device to get draw working.

I also got hung up on why my drawer wasn't working. It was because I was assigning it to a variable in a function rather than on the controller. This caused it go to out of scope immediately after it was created. I added it as property of ViewController and "draw" started getting called.