Metal003 Find the time since the last draw
Find the time since the last draw so I can know how much to update the world each time draw is called. Out of curiosity I put a bit that prints the preferred FPS and what the delta should be based on that. They are pretty close to the same but I assume that preferred FPS won't take it to account real lag caused by overloaded CPU or GPU. I'm going to stick with CACurrentMediaTime for now.