Metal003 I got a the green dot to move!
The green dot moves because it's translated in the GameWorld the result of which is handed off to be rendered. I had to expose some internals of Vertices and Point that I may want to reconsider so that GameWorld can decide if they've moved too far. I may want to reconsider that.

I should remove the vertices being passed to the Drawer as a separate set of vertices and rely on those coming from the World. I also need to figure out how to have the World track many objects and hand them off to the GPU to be rendered.

Once I have a cube or something in here I'll want to only track one point(center?), translate that in the world, then provide a translation matrix that the GPU can use to translate the rest of the points. Otherwise, the CPU ends up doing all the work that the GPU should be doing.