Metal005 Use a static function to create a world
I want to hide the type of world being returned so that all knowledge of it is through the protocol making it easier to quietly replace it with something else that conforms to the protocol. You can't call methods on protocols even through extensions in Swift. I saw the example of MTLCreateSystemDefaultDevice() which also hides the exact type of MTLDevice returned. This seemed like useful and idiomatic way to hide the implementation class so I'm giving that a shot.