Metal005 Stop using the defaults and remove transformation from init
The original idea behind the defaults was to serve as a shim while refactoring plus the values, especially rate, seem even more magical when seen as a value inherent to a GMNode rather than a property inherent to a GMWorld.

I moved transformation out of init for now because I don't yet have a reason to have it be nothing other than the identity matrix for now.