Metal005 Use tangent to calculate perspective projection
When I was refactoring it occurred to me that something didn't seem right about the way I calculated the perspective projection. Especially compare to the ways I saw other people doing it. I spent a bit of time staring at a triangle and a frustum or two. I say SOH-CAH-TOA a few times . Then I realized that the z coordinate is always the adjacent and I know the angle from the FOV angle so I can use that to find the opposite! Amazing!