Remove parent for now.
If I remove the parent I'll need to find some other way to bubble up events from nodes. I may be able to keep track of the parents as I traverse the graph down to the node that fires the event, fire the event, and roll back up the graph using a list of parents, firing events on their children as I climb, then if one of children needs to mutate, creating a new graph, and continuing to move up the tree.

It should be interesting.