Pass the width and height of the screen to the world.
I want the world to be able to make a decision about what button on the screen was pressed. This is an initial implementation of the button. Once I get it working on different portions of the screen I want to try doing a collision test between a ray and a sphere.