Switch to Float rather than CGFloat where ever possible.
I might regret this because I don't know enough about the difference between Float and CGFloat. What I do know is that I need to Float for SIMD matrices. By restricting it I should have a better of understanding of why I need CGFloat when I bonk my head on a problem that requires it.