Implement DeleteNode on GMImmutableContainerNode
In some ways this went how I want it to and in others it didn't. I had to remove the implementation of GMNode for now. I wasn't able to return enough type information to delete the nodes as GMNodes. I think I need to rework GMNode a bit to make that possible. In order to not bite off too much in a single bit of work I removed GMNode protocol from GMImmutableContainerNode.